Fix bluesky error when calling ALT without vspd
ALT always calls set_cleared_fl with vspd in kwargs (vspd=None if it is not in request args). The set_cleared_fl() method expected vspd was either set or not provided. This led to BlueSky syntax error when ALT was called without vspd (could not parse NONE as vspd). This has now been fixed.