Remove hard-coded build tags
To test locally, images were build against specific hashes for sanity
checks. If this same labelled images it not present on another users system,
launches will fail.

It is felt best to remove this information in the docker-compose file
such that anyone can built with the version specified by others.

	modified:   tests/integration/machcoll/docker-compose.yml